Moving to qa_endpoints
